added logic to session management
