+++ frontend/pages/home.html
@@ -7,9 +7,6 @@
             <a href="launchGame">Launch a game</a>
         </li>
         <li>
-            <a href="connectToGame">Connect to a game</a>
-        </li>
-        <li>
             <a href="about">About</a>
         </li>
     </ul>
